+++ commands/digitalungdom/memberlink.js
@@ -8,6 +8,6 @@ module.exports = {
   serverOnly: false,
   adminOnly: false,
   execute( message, args ) {
-    message.reply( 'här är länken att bli medlem: https://digitalungdom.se/bli-medlem' );
+    message.reply( 'här är länken att bli medlem: https://digitalungdom.se/bli-medlem' ).then( msg => { msg.delete( 10000 ); } );
   },
 };