+++ commands/digitalungdom/report.js
@@ -17,18 +17,18 @@ module.exports = {
     let reportedUser;
 
     if ( message.channel.type === 'text' ) {
-      if ( !message.mentions.members.first() ) return message.reply( 'du måste @ vem du vill reporta.' );
-      if ( args.length < 2 ) return message.reply( 'du måste ge en kort anledning.' );
+      if ( !message.mentions.members.first() ) return message.reply( 'du måste @ vem du vill reporta.' ).then( msg => { msg.delete( 10000 ); } );
+      if ( args.length < 2 ) return message.reply( 'du måste ge en kort anledning.' ).then( msg => { msg.delete( 10000 ); } );
 
       reportedUser = message.mentions.members.first().user;
-    } else if ( message.channel.type === 'dm' ) {
-      if ( args.length < 4 ) return message.reply( 'du måste ange vem du vill reporta och ge en kort anledning.' );
+    } else {
+      if ( args.length < 4 ) return message.reply( 'du måste ange vem du vill reporta och ge en kort anledning.' ).then( msg => { msg.delete( 10000 ); } );
       let reported = args[ 0 ];
       if ( reported.startsWith( '@' ) ) reported = reported.slice( 1 );
       if ( /#\d{4}$/.test( reported ) ) reported = reported.slice( 0, -5 );
 
       reportedUser = client.users.find( user => user.username == reported );
-      if ( !reportedUser ) return message.reply( 'kunde inte hitta användare, se till så att det inte är deras nickname och att alla versaler är korrekta.' );
+      if ( !reportedUser ) return message.reply( 'kunde inte hitta användare, se till så att det inte är deras nickname och att alla versaler är korrekta.' ).then( msg => { msg.delete( 10000 ); } );
     }
 
     let reason = args;
@@ -70,6 +70,6 @@ module.exports = {
     const notificationChannel = guild.channels.find( ch => ch.name === 'notifications' );
     notificationChannel.send( '@here, ny notifikation', { 'embed': notification } );
 
-    return message.author.send( 'Tack för din medverkan, vi kommer inom en snar framtid att granska din anmälan.' );
+    return message.author.send( 'Tack för din medverkan, vi kommer inom en snar framtid att granska din anmälan.' ).then( msg => { msg.delete( 10000 ); } );
   },
 };