+++ CHANGELOG.md
@@ -1,1 +1,13 @@
 # 1.1.0
+
+## new commands
+
+-   kick/ban and logging to database
+-   dubstatus, an admin command to change the status of the bot
+
+## improvements
+
+-   remove no access commands from help page (e.g. admin commands)
+-   the bot can now only be called from the bot channels (if the bot was called from the server)
+
+## deprecations